Include provenance into generated OSGI artifacts
provenance classes are used in the public API for CsarImporter.
Any consumers of CsarImporter must also have access to
provenance.exceptions. This also applies for OSGI consumers!

This commit also preemptively configures the provenance module
for deployment to the internal repository to sidestep future issues
with maven-projects depending on the artifact.

Signed-off-by: Clemens Lieb <vogel612@gmx.de>