Store Provenance State in Swarm
When exporting a provenance CSAR, all new or edited files since the last
export of the same servicetemplate are stored in the Ethereum Swarm
network.
Furthermore, these files are retrievable on-demand so we can visually show changed
content between any two exported CSARs of the same servicetemplate.
Accountability UI is enhanced and a configuration tab for it is added.

Signed-off-by: Ghareeb Falazi <ghareeb.falazi@hotmail.com>